+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.1 - 2017-04-26
+
+### Added
+- nyc marathon bot
+
 ## 0.1 - 2017-04-25
 
 ### Tested